Replaced usage of delta: Float in Workflow with delta: Double for better precision and consistency.

+++ core/src/main/scala/org/sgine/workflow/Workflow.scala
@@ -33,7 +33,7 @@ class Workflow(val items: List[WorkflowItem]) extends WorkflowItem with Updatabl
     current = null
   }
 
-  def act(delta: Float) = {
+  def act(delta: Double) = {
     if (current == null && !currentItems.isEmpty) {
       current = currentItems.head
       current.begin()
